requirements.txt
fastapi
fastapi-mail
confluent-kafka
python-dotenv
uvicorn
nest_asyncio